pyproject.toml
[project]
name = "running-coach"
version = "0.1.0"
description = "RunWhisperer — AI running coach OpenClaw skill backed by Strava + Power of 10"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8",
    "pytest-cov>=6",
    "mypy>=1.14",
    # Script deps needed for testing
    "fire>=0.7",
    "requests>=2.32",
    "python-dotenv>=1.0",
    "beautifulsoup4>=4.12",
    "python-telegram-bot>=21.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
